pyproject.toml
[project]
name = "nkt-easeme-api"
version = "0.1.0"
description = "EaseMe FastAPI backend (Supabase + OpenAI)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "httpx>=0.27",
  "pyjwt[crypto]>=2.9",
  "python-jose[cryptography]>=3.3",
  "SQLAlchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "orjson>=3.10",
  "tenacity>=9.0",
]

[tool.uvicorn]
factory = true
host = "0.0.0.0"
port = 8000
